% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{algorithm}
\alias{algorithm}
\title{Extract algorithmic options from Blimp output}
\usage{
algorithm(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing algorithmic options
}
\description{
Extracts the algorithmic options section from Blimp output.
}
