% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{as.data.frame,blimp_obj-method}
\alias{as.data.frame,blimp_obj-method}
\title{Convert blimp_obj to data.frame}
\usage{
\S4method{as.data.frame}{blimp_obj}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{blimp_obj} object}

\item{row.names}{NULL or a character vector giving row names}

\item{optional}{Logical. If TRUE, setting row names is optional}

\item{...}{Additional arguments passed to as.data.frame}
}
\description{
Convert blimp_obj to data.frame
}
