% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{as.matrix,blimp_obj-method}
\alias{as.matrix,blimp_obj-method}
\title{Convert blimp_obj to matrix}
\usage{
\S4method{as.matrix}{blimp_obj}(x, ...)
}
\arguments{
\item{x}{A \code{blimp_obj} object}

\item{...}{Additional arguments (unused)}
}
\value{
A numeric matrix containing the MCMC iterations from the model.
}
\description{
Convert blimp_obj to matrix
}
