% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{as.mitml}
\alias{as.mitml}
\title{Coerces a \code{\link{blimp_obj}} or \code{blimp_bygroup} to a \code{mitml.list}}
\usage{
as.mitml(object)
}
\arguments{
\item{object}{\code{\link{blimp_obj}} or \code{blimp_bygroup} object}
}
\value{
A list of class \code{"mitml.list"} containing the imputed data sets,
suitable for use with functions from the 'mitml' package.
}
\description{
Coerces a \code{\link{blimp_obj}} or \code{blimp_bygroup} to a \code{mitml.list}
}
