% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\docType{class}
\name{blimp_cp-class}
\alias{blimp_cp-class}
\title{S4 class for Blimp model comparison results}
\description{
Result object from comparing two Blimp models.
}
\section{Slots}{

\describe{
\item{\code{table}}{Matrix of comparison results}

\item{\code{model0}}{The reference blimp_obj model}

\item{\code{model}}{The comparison blimp_obj model}

\item{\code{greaterThan}}{Logical indicating direction of comparison}
}}

