% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare,blimp_obj,blimp_obj-method}
\title{Compare two Blimp models}
\usage{
compare(
  model0,
  model,
  use = "mean",
  greaterThan = TRUE,
  suffixes = c("R2: Coefficients", "R2: Level-2 Random Intercepts",
    "R2: Level-2 Random Slopes", "R2: Level-3 Random Slopes",
    "R2: Level-3 Random Intercepts", "R2: Residual Variation",
    "R2: Level-1 Residual Variation")
)

\S4method{compare}{blimp_obj,blimp_obj}(
  model0,
  model,
  use = "mean",
  greaterThan = TRUE,
  suffixes = c("R2: Coefficients", "R2: Level-2 Random Intercepts",
    "R2: Level-2 Random Slopes", "R2: Level-3 Random Slopes",
    "R2: Level-3 Random Intercepts", "R2: Residual Variation",
    "R2: Level-1 Residual Variation")
)
}
\arguments{
\item{model0}{Reference model used to get the cutpoint}

\item{model}{Comparison model to evaluate}

\item{use}{A character, numeric value, function, or list. If character, recognizes
'mean' and 'median'. If numeric < 1, acts as proportion for quantile. If numeric >= 1,
converts to proportion out of 100. If function, applies that function. Use list for
multiple columns.}

\item{greaterThan}{Logical. If TRUE, evaluates proportion greater than cutpoint}

\item{suffixes}{Character vector of parameter name suffixes to compare. Defaults to
all R-squared values.}
}
\value{
A \code{blimp_cp} object containing comparison results
}
\description{
Generates a blimp_cp object to compare two models by evaluating how many
iterations are above/below a cutpoint.
}
\section{Methods (by class)}{
\itemize{
\item \code{compare(model0 = blimp_obj, model = blimp_obj)}: Compare two blimp_obj models

}}
\note{
Due to R restrictions, lists of functions will not give useful printed names.
}
