% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{datainfo}
\alias{datainfo}
\title{Extract data information from Blimp output}
\usage{
datainfo(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing data information
}
\description{
Extracts the data information section from Blimp output.
}
