% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{estimates}
\alias{estimates}
\title{Obtain MODEL ESTIMATES from \code{blimp_obj} or \code{blimp_out}}
\usage{
estimates(object)
}
\description{
Obtain MODEL ESTIMATES from \code{blimp_obj} or \code{blimp_out}
}
