% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{modelfit}
\alias{modelfit}
\title{Extract model fit from Blimp output}
\usage{
modelfit(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing model fit information
}
\description{
Extracts the model fit section from Blimp output.
}
