% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{modelinfo}
\alias{modelinfo}
\title{Extract MODEL INFORMATION from \code{blimp_obj} or \code{blimp_out}}
\usage{
modelinfo(object)
}
\description{
Extract MODEL INFORMATION from \code{blimp_obj} or \code{blimp_out}
}
