% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{output}
\alias{output}
\title{Extract output from blimp_obj}
\usage{
output(object)
}
\arguments{
\item{object}{A \code{blimp_obj} object}
}
\value{
A \code{blimp_out} object containing the raw Blimp output text
}
\description{
Extracts the raw Blimp output text from a blimp_obj.
}
