% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{plot,blimp_obj-method}
\alias{plot,blimp_obj-method}
\title{Create Traceplots for \code{blimp_obj}}
\usage{
\S4method{plot}{blimp_obj}(x, y, colors = NULL, ...)
}
\description{
Create Traceplots for \code{blimp_obj}
}
