% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblimp_syntax.R
\name{print.blimp_syntax}
\alias{print.blimp_syntax}
\title{Print \code{blimp_syntax}}
\usage{
\method{print}{blimp_syntax}(x, ...)
}
\arguments{
\item{x}{A \code{blimp_syntax} object}

\item{...}{Additional arguments (unused)}
}
\value{
The \code{blimp_syntax} object (invisibly). Called for its side effect
of printing the formatted syntax to the console.
}
\description{
Print \code{blimp_syntax}
}
