% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{psr}
\alias{psr}
\title{Extract POTENTIAL SCALE REDUCTION output from \code{blimp_obj} or \code{blimp_out}}
\usage{
psr(object)
}
\description{
Extract POTENTIAL SCALE REDUCTION output from \code{blimp_obj} or \code{blimp_out}
}
