% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_syntax.R
\name{rblimp_syntax}
\alias{rblimp_syntax}
\title{Generates a syntax 'imp' file for blimp based on rblimp}
\usage{
rblimp_syntax(
  model,
  data,
  burn,
  iter,
  seed,
  thin,
  nimps,
  latent,
  randomeffect,
  parameters,
  clusterid,
  ordinal,
  nominal,
  transform,
  fixed,
  center,
  chains,
  simple,
  waldtest,
  options,
  output
)
}
\description{
Generates a syntax 'imp' file for blimp based on rblimp
}
