% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{resid,blimp_obj-method}
\alias{resid,blimp_obj-method}
\title{Residuals scores from \code{blimp_obj}}
\usage{
\S4method{resid}{blimp_obj}(object, ...)
}
\arguments{
\item{object}{A \code{blimp_obj} object}

\item{...}{Additional arguments passed to residuals}
}
\value{
A list of data frames, one per imputation, each containing the
residual columns from the model.
}
\description{
Residuals scores from \code{blimp_obj}
}
