% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{show,blimp_cp-method}
\alias{show,blimp_cp-method}
\title{Show method for blimp_cp}
\usage{
\S4method{show}{blimp_cp}(object)
}
\arguments{
\item{object}{A \code{blimp_cp} object}
}
\description{
Show method for blimp_cp
}
