% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{show,blimp_obj-method}
\alias{show,blimp_obj-method}
\title{Show method for blimp_obj}
\usage{
\S4method{show}{blimp_obj}(object)
}
\arguments{
\item{object}{A \code{blimp_obj} object}
}
\value{
No return value, called for the side effect of printing the
model estimates to the console.
}
\description{
Show method for blimp_obj
}
