% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{with,blimp_bygroup-method}
\alias{with,blimp_bygroup-method}
\title{Fit Model across imputations with \code{mitml} package using \code{\link{by_group}}}
\usage{
\S4method{with}{blimp_bygroup}(data, expr, ...)
}
\description{
Fit Model across imputations with \code{mitml} package using \code{\link{by_group}}
}
