% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{write.blimp}
\alias{write.blimp}
\alias{write.blimp,blimp_syntax-method}
\alias{write.blimp,blimp_out-method}
\alias{write.blimp,blimp_obj-method}
\title{A function to write out blimp input and output from a model}
\usage{
write.blimp(object, folder = "")

\S4method{write.blimp}{blimp_syntax}(object, folder = "")

\S4method{write.blimp}{blimp_out}(object, folder = "")

\S4method{write.blimp}{blimp_obj}(object, folder = "")
}
\arguments{
\item{object}{A \code{\link{blimp_obj}}.}

\item{folder}{a location to a folder to write input and output}
}
\description{
A function to write out blimp input and output from a model
}
\section{Methods (by class)}{
\itemize{
\item \code{write.blimp(blimp_syntax)}: Write blimp_syntax to file

\item \code{write.blimp(blimp_out)}: Write blimp_out to file

\item \code{write.blimp(blimp_obj)}: Write blimp_obj files to folder

}}
\examples{
\dontshow{if (has_blimp()) withAutoprint(\{ # examplesIf}
# Generate Data
mydata <- data.frame(x = rnorm(1000), y = rnorm(1000))

# Nonsensical model
mdl <- rblimp(
    c(
        'y <- x*2',
        'f1 <- 1'
    ),
    mydata,
    seed = 3927,
    nimps = 2,
    latent = ~ f1,
    center = cgm ~ x,
    fixed = ~ x
)

# Write out input and output
\dontrun{
write.blimp(mdl, "folder_location")
}
\dontshow{\}) # examplesIf}
}
