% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{algorithm}
\alias{algorithm}
\title{Extract ALGORITHMIC OPTIONS output from \code{blimp_obj} or \code{blimp_out}}
\usage{
algorithm(object)
}
\description{
Extract ALGORITHMIC OPTIONS output from \code{blimp_obj} or \code{blimp_out}
}
