% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblimp_syntax.R
\name{as.character.blimp_syntax}
\alias{as.character.blimp_syntax}
\title{Convert \code{blimp_syntax} to character vector}
\usage{
\method{as.character}{blimp_syntax}(x, ...)
}
\arguments{
\item{x}{A \code{blimp_syntax} object}

\item{...}{Additional arguments (unused)}
}
\value{
A character string containing the formatted 'Blimp' input syntax
with a header comment and all commands.
}
\description{
Convert \code{blimp_syntax} to character vector
}
