% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_syntax.R
\name{as.character.blimp_syntax}
\alias{as.character.blimp_syntax}
\title{Convert `blimp_syntax` to character vector}
\usage{
\method{as.character}{blimp_syntax}(x, ...)
}
\description{
Convert `blimp_syntax` to character vector
}
