% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\docType{class}
\name{blimp_obj}
\alias{blimp_obj}
\alias{blimp_obj-class}
\title{S4 class for Blimp model results}
\description{
The main result object containing Blimp model estimates, iterations, and output.
}
\section{Slots}{

\describe{
\item{\code{call}}{The function call that created the object}

\item{\code{estimates}}{Matrix of parameter estimates}

\item{\code{burn}}{List of burn-in information for each chain}

\item{\code{iterations}}{Data frame of MCMC iterations}

\item{\code{psr}}{Data frame of Potential Scale Reduction values}

\item{\code{imputations}}{List of imputed datasets}

\item{\code{average_imp}}{Data frame of average imputations}

\item{\code{variance_imp}}{Data frame of imputation variances}

\item{\code{waldtest}}{Data frame of Wald test results}

\item{\code{simple}}{Data frame for simple slopes analysis}

\item{\code{syntax}}{The blimp_syntax object used to run the model}

\item{\code{output}}{The blimp_out object containing raw output text}
}}

