% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jn_plot.R
\name{compute_condeff}
\alias{compute_condeff}
\title{Convenience Function for computing conditional effects for \code{\link{jn_plot_func}}}
\usage{
compute_condeff(value1, value2)
}
\arguments{
\item{value1}{The base value}

\item{value2}{The value to change as a function of moderator}
}
\value{
a \code{\link{function}}
}
\description{
Convenience Function for computing conditional effects for \code{\link{jn_plot_func}}
}
\seealso{
\code{\link[=jn_plot_func]{jn_plot_func()}}
}
