% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{datainfo}
\alias{datainfo}
\title{Extract DATA INFORMATION output from \code{blimp_obj} or \code{blimp_out}}
\usage{
datainfo(object)
}
\description{
Extract DATA INFORMATION output from \code{blimp_obj} or \code{blimp_out}
}
