% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\docType{class}
\name{describe-class}
\alias{describe-class}
\alias{show,describe-method}
\title{The 'describe' class for object summaries}
\usage{
\S4method{show}{describe}(object)
}
\arguments{
\item{object}{An object of class \code{describe}.}
}
\description{
An S4 class that holds a summary data frame and formatting
information. It extends the \code{data.frame} class and is the return type for
the \code{\link[=describe]{describe()}} function.
}
\section{Slots}{

\describe{
\item{\code{digits}}{A numeric value indicating the number of digits to display.}
}}

