% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{estimates}
\alias{estimates}
\title{Extract model estimates from Blimp output}
\usage{
estimates(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing model estimates
}
\description{
Extracts the model estimates section from Blimp output.
}
