% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{modelinfo}
\alias{modelinfo}
\title{Extract model information from Blimp output}
\usage{
modelinfo(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing model information
}
\description{
Extracts the model information section from Blimp output.
}
