% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{names,blimp_obj-method}
\alias{names,blimp_obj-method}
\title{Return variable names from \code{blimp_obj} object}
\usage{
\S4method{names}{blimp_obj}(x)
}
\arguments{
\item{x}{A \code{blimp_obj} object}
}
\value{
A character vector of variable names from the average imputation
data.
}
\description{
Return variable names from \code{blimp_obj} object
}
