% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{predict,blimp_obj-method}
\alias{predict,blimp_obj-method}
\title{Predicted scores from \code{blimp_obj}}
\usage{
\S4method{predict}{blimp_obj}(object, ...)
}
\arguments{
\item{object}{A \code{blimp_obj} object}

\item{...}{Additional arguments (unused)}
}
\description{
Predicted scores from \code{blimp_obj}
}
