% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{psr}
\alias{psr}
\title{Extract Potential Scale Reduction (PSR) output}
\usage{
psr(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{blimp_out} object containing PSR output
}
\description{
Extracts the PSR convergence diagnostic section from Blimp output.
}
