% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{residuals,blimp_obj-method}
\alias{residuals,blimp_obj-method}
\title{Residuals scores from \code{blimp_obj}}
\usage{
\S4method{residuals}{blimp_obj}(object, ...)
}
\description{
Residuals scores from \code{blimp_obj}
}
