% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{standardized}
\alias{standardized}
\title{Extract standardized solutions from Blimp}
\usage{
standardized(object)
}
\arguments{
\item{object}{A \code{blimp_obj} or \code{blimp_out} object}
}
\value{
A \code{\link[base:matrix]{base::matrix}} with standardized solutions
}
\description{
Extracts the data information section from Blimp output.
}
