% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_blimp.R
\name{update_blimp}
\alias{update_blimp}
\title{Check for Update of Blimp}
\usage{
update_blimp()
}
\description{
Check for Blimp updates
}
\details{
This function is called on loading the package.
If the session is interactive then it will not check for an update.

Checking update on load can be disabled by setting options \code{check_blimp_update} to \code{FALSE}. See example below.
}
\examples{
# Disable Update checking on load
options(check_blimp_update = FALSE)
}
