% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_blimp.R
\name{update_blimp}
\alias{update_blimp}
\title{Runs the Blimp Updater in order to Update Blimp}
\usage{
update_blimp()
}
\description{
Runs the Blimp Updater in order to Update Blimp
}
\details{
If the session is not an interactive session then it will not open the updater.
}
