% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_group.R
\name{with,blimp_bygroup-method}
\alias{with,blimp_bygroup-method}
\title{Fit Model across imputations with \code{mitml} package using \code{\link{by_group}}}
\usage{
\S4method{with}{blimp_bygroup}(data, expr, ...)
}
\arguments{
\item{data}{A \code{blimp_bygroup} object}

\item{expr}{An expression to evaluate on each imputation}

\item{...}{Additional arguments (unused)}
}
\value{
A list of class \code{"mitml.result"} containing the results of
evaluating \code{expr} on each imputed data set.
}
\description{
Fit Model across imputations with \code{mitml} package using \code{\link{by_group}}
}
