% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{with,blimp_obj-method}
\alias{with,blimp_obj-method}
\title{Fit Model across imputations with \code{mitml} package}
\usage{
\S4method{with}{blimp_obj}(data, expr, ...)
}
\arguments{
\item{data}{A \code{blimp_obj} object}

\item{expr}{An expression to evaluate on each imputation}

\item{...}{Additional arguments (unused)}
}
\description{
Fit Model across imputations with \code{mitml} package
}
