% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blimp_obj.R
\name{write.blimp}
\alias{write.blimp}
\title{A function to write out blimp input and output from a model}
\usage{
write.blimp(object, folder = "")
}
\arguments{
\item{object}{A \code{\link{blimp_obj}}.}

\item{folder}{a location to a folder to write input and output}
}
\description{
A function to write out blimp input and output from a model
}
\examples{
\dontshow{if (has_blimp()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Generate Data
mydata <- data.frame(x = rnorm(1000), y = rnorm(1000))

# Nonsensical model
mdl <- rblimp(
    c(
        'y <- x*2',
        'f1 <- 1'
    ),
    mydata,
    seed = 3927,
    nimps = 2,
    latent = ~ f1,
    center = cgm ~ x,
    fixed = ~ x
)

# Write out input and output
\dontrun{
write.blimp(mdl, "folder_location")
}
\dontshow{\}) # examplesIf}
}
